add new module basic structure.
add new field replacement_product_id to the product.template model and view.

The Status of the product now is manage in the header of the product form view as a clickable statusbar widget and with default value `In Development`.

some search filters were added to search and group by products lifecycle status.

Add a rule to restrict the edit a product. Only the product manager can edit the product and the product variant.

new wizard that extract the replacement products of a discontinued product. Show then and let the user to select one.

add the documentation of the module in description index.html file and update the README file.

do not let to create a a purchase order with a discontinued product. will reaise a warning indicating that the product is discontinued.

add new logic in the purchase order lines. When the user add a discontinued product will indicate a warning with this information and will require to the user to select a replacement product.

add check the purchase order line to look for discontinued products in the lines, if is the case will raise a warning indicating that the purchase order can not have purchase order lines with discontinued products.

the functionality was moved from the product.template model to the product.product model.

add label to identificate that the replacement product are for purchase operations.

add demo data for product states and its replacement products.

only show as a replacement the products that are not obsolete and active

add sequence to the purchase order and show the line sequence in the message of discontinued product.

Multiple Fixes:

[pylint]
- W0622 redefinition of Warning import from openerp.

[flake8]
- E126 continuation line over-indented for hanging indent
- E231 missing whitespace after ','

hide the status state in the product template. this field is not used anymore. the new state2 state defined in product.product is what defines the status of the variant.

show the status of every variant in the product.template form at the Inventory > Status section. This is only show if watching from the product template form but not form the product variant form.

i[ADD] Replacement Product in the purchase order line was deleted to use a new field Discontinued Product. This new field holds the obsolete product and when is fill will add a domain into the Product field with the good replacements products (active and not obsolete products marked as a replacements of the current discontinued product). In other case, if the user try to fill the Product field with a discontinued product will show a warning indicating that can not use a discontinued product in the purchase order and will move this value to the Discontinued Product field, cleaning the Product field so you can select a replacement product.

In a technical matters, some methods were modularizate to optimize code, the
methods docstring were updated to be more detail and the old comments in code
were deleted.

show the display name of the variant instead of the code.

update description of the module.

change the replacement wizard product to be call from the purchase order and extract the information of the discontinued products in the order lines to be able to select the replacements configured in the product. The purchase orders will be updated will move the product to the discontinued product field and will add the new replacement selected in the wizard to the product field.

new module product_lifecycle translations. Some fixes to the code was made.

update the module help indicating how it works the Replace Discontinued Product Wizard.

the new state2 field in the product.product module now have the copy=False attribute to make sure than when make a copy do not copy the field just add the default.

Improves over the Replacement Wizard: - [ADD] Add constraint to do not let to save a replacement line if the discontinued product is not a obsolete product or if the replacement product is a obsolete product. - [FIX] resolve pylint error about importing openerp.exceptions.Warning - [CC] reorganizate the imports taking into accounv v8 style.

update the demo data to construct better unit test cases.

add unti test to the product_lifecycle module.

TODO Comment testing related to demo data because changes made have been commited to the database. This generate a unit test error when running test a second time. This was report to odoo. after is fixes uncomment the section of code.

Fix importing warnings.

resolve problem with the unit test. The test was not doing a rollback to the end of the test so when trying to run the test after the first time the database was corrupted and the test return errors. Change the schema of the test from a SingleTransactionCase to a TransactionCase. Now every test method is separate. Some changes to modularizate the test were made to fullfile the invidivuality per test method. Also the not functional commented test were uncommented and are working propertly.

when no order lines muts return a empty list not a boolean value. this was generating and error in the test of yoytec_customer_rma_workflow module.