[ADD][sale_block_cancellation] Module to block cancellation of sales with validation added
Added a new button in Sale Orders to only allow cancel a SO when all the
pickings related with this are on ready to transfer or in
waiting for another move state, or even with pickings on different
states if the moved product quantities are correctly returned.

The original button to cancel is added to a new group Force Sale Cancel
to can allow cancel a Sale Order out of these cases.