[FIX] sale_team_warehouse: monkey patch is a bad approach.
Avoiding the usage of monkey patching flaky tests appear when it is
done.

With this refactor 2 things need to be considered:

1. To use the feature of set the default value of warehouse in a model we
need to inherit (ala mail.thread) the model where the field warehouse_id
will be filled if it exists.

2. If the model default.warehouse is inherited and the field is not
setted properly as Many2one related to warehouse model, then the
inheritance will be simply ignored.

Read the test cases to understand the usage.

Is closes #694