[REF] product_available_by_warehouse: remove domain route in show message
When it was used 'show message' field to see product available the
domain route should consider only the routes that are available for
sale order line.

Also, swapped positions between show_message field and product_id
field. For usability is better to choose first the product.

In order to solve issue #1054