[REF] stock_landed_costs_average: Separate landed cost lines creation
We are separating creation of Landed Lost Line objects to be able to
adapt this object creation inside other modules.

Now if you want to change values of the landed cost line being created
you only need to overwrite this method and return the updated
dictionary.

This changes were proposed by @hbto to be used by other Vauxoo's projects.