[ADD] sale_order_shipping_date: added compute field shipping_date
This module add a field into Sale Order that show the Date
by which the are sure to be delivered. In contrast to commitment
date, not only is taken into account the product lead time, but
also the resupply delay, the also the resupply delay, the
delivery lead time from supplier and the manufacturing delay.

Technically, the shipping date field comes to replace the
commitment date.