[MIG] partner_credit_limit and payment_term_type: Migrated to v9 (#1076)
* [IMP] partner_credit_limit, payment_term_type: Change license and version

* [IMP] Is eliminated reference to sales workflow, in version 9.0 the sale
workflow is deleted. It moves the call of the function check_limit to the
function action_confirm

* [IMP] payment_term_type: modified reference the field
module_product_email_template, because this does not exist in version
9.0, change to module_account_batch_deposit

* [FIX] payment_term_type: Fixed field value malformed string in demo data

* [IMP] partner_credit_limit: Modified fields are renamed or no longer
exist in Odoo 9

* [IMP] partner_credit_limit: Modified heritage res_partner view, to add
the credit_limit field, which is created by Odoo, but is not added to
any view. Also modified the action that opens the journal items overdue payment

* [FIX] partner_credit_limit: Changed file name account_invoice, and
corrected function warning message.

* [IMP] partner_credit_limit: Test adapted to work in Odoo 9. Created
common file to add function to load the demo data of account module

* [IMP] payment_term_type: It is eliminated creating lines of balance
payment term type, because in Odoo 9 are created by default

* [IMP] partner_credit_limit: Add contributor

* [FIX] partner_credit_limit: translation creadit to credit is changed