[IMP] account_move_line_base_tax: Removed amount validation to diot movements
In this module was added a validation to verify that movements that will
be considered in DIOT report must have amount in the credit. But to invoice
refunds the logic change, and the validation is incorrect.

The DIOT report already have the validation to movements with amount 0,
if this are created manually. By this reason is removed the validation
here.