[FIX] Consumable and service type products are excluded in qty2produce (#1345)
Consumable and service type products are removed from the bill of materials when are analyzed in the get_qty_availble_to_produce method.

This bug was discovered in the task#31810 of the ticket#5182.

The client tries to produce a product with bills of material that include service, so the function doesn't consider the service because this doesn't appear in the order, and the function always returns 0.