[IMP] stock_by_warehouse[_sale,_purchase]: Fix big performance overhead (#1358)
We need avoid using compute fields as possible because of the following issue:
 - https://github.com/odoo/odoo/issues/30578
warehouses_stock is a computed field used just one time but re-computed too many times:
 - Opening sale order (one time by each line)
 - Saving sale order
 - Choosing a product_id
And it computation is more and more slow when the database grows.
It is making a big overhead saving sale.order.

With this change we have the same behaviour using it just for onchange
product_id or forcing from a seudo-button (field boolean)
A real button is not used since that it requires save the record to work

Using the technical of fields with store=False and onchange methods

Check more details from https://github.com/Vauxoo/addons-vauxoo/pull/1358