[ADD] New feature sale price usd
New 'Sale Price in USD' field on the product's form.
The Sale Price of Odoo is calculated in base to Sale Price in USD daily with the last rate of dollar.
Allow computation in pricelist based on Sale Price in USD.
The sales price is calculated daily with the new rate
The sales price is re-calculated if Sales Price in USD field is changed or setted.
The Sales Price in USD is re-calculated if sales price is changed or setted.