[IMP] partner_credit_limit: Improve the module for multicompany (#1371)
- For instances of multiple companies, the credit limit value or days of grace
does not differentiate between company, so if the customer has a credit limit
`100` with `company A`, that same value will be in the `company B`.

In some cases these values changes according to the company.

With this change the `account.move.line` are searched according to the
company and additionally the credit limit fields and grace days now
depend on the company.

- Migration script to set the old values for the `credit_limit` and
`grace_payment_days` fields.