[ADD] internal_transfer_with_agreed_amount: New module (#1393)
- When foreign currencies are bought, in some cases an amount is agreed
for the transaction and it is not exactly that resulting from the
official exchange rate.

- When payments are made or received in foreign currency and the customer
or payment service uses its own exchange rate.

For these cases this module allows an internal transfer between
different currencies, setting an agreed amount.

This agreed amount will be used to set the amount currency in the
journal entries.

A new module is created to avoid duplication of code in different
clients that require this process.