[FIX] partner_credit_limit: Add hook to initialize credit limit (#1395)
Starting from 1e18a7457, the partner's credit limit field is
company-dependent, so their values need to be set manually, because they
are properties instead of a standard field.

A migration script was provided on the aforementioned commit to set that
value when the module is already installed. However, since that field is
a native one, values need to be set also when the module is installed.

This commit fixes the above by implementing a post-init hook to set
required values.