[ADD] stock_by_warehouse: Product availability in locations Widget
Adding a new section for the widget `warehouse` that will depend on an
option called `by_location`, if is set to True it will allow the user
to find all the locations where the product (`product.product`) is
available, and the quantity that is in stock.
The widget it will allow the user to see the locations and its
quantity for each warehouse.
And updating the image for the widget by warehouse on the README file.