[ADD] Stock By Warehouse on Stock Module
Adding a new module called `stock_by_warehouse_stock` where the widget
`warehouse` (with the option `by_location` set on True) is going to be used
from the `stock.move.operations` view, to allow the users to see where the
product is located on the warehouse of the stock move.
Also, adding a new field called `suggested_location_id` where the is
going to be set the location where the product has the highest quantity on
stock. For later using it as a default location in a new field `location_id`
for the stock move lines.

This widget will only be seen on the stock moves when its transfer has
the type of outgoing, and the default location will be set when the
stock move doesn't have the state of 'done' or 'cancel'.