[ADD] stock_custom_exchange_rate_date: Module to set custom rate date
This module allows to set a customized date on transfers to be used when
computing product unit prices, instead of the date when the transfer  is
validated, in case exchange rates are involved.

The use for the above is that created journal entries take the newly
computed unit price, which makes the inventory to be valuated using the
specified rate.