[FIX] mrp_partial_production: Missing quant reservation when producing less than the qty reserved
Before this commit when you recorded production less than the qty you
reserved on raw materials, the stock quants did not get reserved with
the remaining qty from previous reservation but inventory lines.

This causes a conflict when trying to cancel/unreserve the mrp
production getting following error:

It is not possible to unreserve more products of %s than you have in stock.

After this commit, now quants reservation it's consistent with inventory
reservation.