[FIX] stock_by_warehouse_sale: Edit SO line using form view
In previous Odoo versions, the form view was poped up when editing a
sale order line, as long as the option "Manage Product packaging" was
enabled. This is no longer the case [1], so we re-enable it manually,
in order to be able to see current inventory when filling a line.

[1] https://github.com/odoo/odoo/commit/e813487e75f1