[FIX] stock_by_warehouse: Fix unittest error caused by action_confirm
The issue was that the action_confirm takes 2 units of product immediately before action_assign.
On v14.0 the products were taken from picking after action_assign. Using action_confirm and
action_assign consecutively mantains the original behaviour from the test. This explained behaviour
is due to stock.move changes where this model handles action_assign when action_confirm is
executed. Functionality is preserved, but the change in the unittest is necessary to follow new
behaviour.

References:
https://github.com/odoo/odoo/commit/a838f2af683e03cd03ba02532f002ccd5c2642f5,
https://github.com/odoo/odoo/blob/14.0/addons/stock/models/stock_move.py#L1169
https://github.com/odoo/odoo/blob/15.0/addons/stock/models/stock_move.py#L1239