[FIX] account_chatter: missing activities on chatters
To be able to show chatter only to certain users, we:
- Hide chatter for everywhere by removing it
- Re-introduce it only for the specific group

However, the new chatter is missing the activity widget, which was
already present on the removed one.

This atts missing activity widget to journal and journal entries, and
improve accounts by adding it also there.

In addition, journal entry's chatter was being placed on  a wrong
position, before attachments preview instead of after it.