[MIG] search_action_domain: Migrate to v15.0 T#59176
This module lets the user to use server actions as a filters in the
"Favorites" section, in other words, the user might use complex queries
to search the records.

Now, "Create A Filter" is not in action section anymore, it is a button
in the server action's form view to create the filter in the selected
model.