[IMP] mail_channel_notify_tag: Welcome mail_channel_notify_tag
This module re-enables channel_ids from the front end when the channels
are mentioned in the messages and then, they are handled to post the
message in the respective channels and an email to the members of the
channels.

The context 'mentioned_channel_ids' is created to handle the channels in
the inheritances.

channel_ids functionality was removed on:
- https://github.com/odoo/odoo/commit/e5338146
- https://github.com/odoo/odoo/commit/59d6bb23