[FIX] controller_report_xls: Get Odoo Style
When the node had a style like
"background-color: #99CCFF; font-weight: bold;" the style that comes
after the "; " (with the white-space after the ;) was not being found on
the method `css2excel`.
That happened because the `get_odoo_style` returned the styles like:
`{'background-color': '#99CCFF', ' font-weight': 'bold'}` and the
`process_css` didn't find ' font-weight' in its keys.

So now cleaning the values on the `get_odoo_style` by removing the
whitespaces from the beginning and the end of the string, the method
will now return `{'background-color': '#99CCFF', 'font-weight': 'bold'}`
and the `process_css` will find 'font-weight' in its keys.