[FIX] stock_by_warehouse: display stock by warehouse in widget
There was an error when trying to load the stock by warehouse widget
caused by the incorrectly use of `onWillUpdateProps` that was change
in [1], this commit fixes that error and also adapt the stock by
warehouse field to be computed. Finally, some styles were adapted to have
a better display and to use recent version of bootstrap.

[1]: https://github.com/odoo/odoo/commit/d682871a