[FIX] internal_transfer_with_agreed_amount: Fixing AssertionError in internal_transfer_with_agreed_amount test
The test was failing due to an attempt to write to the invisible field
'destination_journal_id'. This occurred because the
'destination_journal_id' field was set to be invisible when the
'is_internal_transfer' field is not True. Consequently, when the
'journal_id' field changed in the form, the 'is_internal_transfer' value
changed to False, triggering the error.

This commit resolves the issue and prevents the AssertionError.