[FIX] default_warehouse_from_sale_team: prevent Sales Team compute on default journal set
Use the cached value of the Sales Team field to access its value as a
record. This avoids computing its value when it has not been computed
yet, which could result in setting the wrong Sales Team on an invoice
form creation. This issue occurs because the company is not yet set, and
its value is required to determine the correct Sales Team [1].

[1]: https://github.com/odoo/odoo/blob/3884a60e/addons/sale/models/account_move.py#L41