[ADD] base_sync_login_email: module to Restrict editing emails of users
It is natively possible for an internal user to edit the e-mail of the
contact associated to their user. This could be a security risk, because
someone could (accidentally?) change their own e-mail to an address
different from the one used to sign up (e.g. an e-mail outside the
organization), which could lead to an external person to receive sensitive
information or intended to be sent only to internal users.

To prevent the above, this new module disables editing e-mails of contacts
associated with internal users if their e-mail matches their user login.
Now, to edit an e-mail, you will need to edit the user login, which will
update contact e-mails automatically, ensuring both login and e-mail are
synced.

This adds two security layers:
- It prevents accidental or unauthorized changes to contact e-mails.
- It restricts e-mail modifications to administrators, as only they have
  permission to edit user records.